requirements.txt
Flask~=3.1.0
pika~=1.3.2
python-dotenv~=1.0.1